Start moving parsing of the document to extra module
